+++ alan_project/alan/urls.py
@@ -7,4 +7,5 @@ urlpatterns = [
     url(r'^scanner/$', views.run_scanner, name='scanner'),
     url(r'^parser/$', views.run_parser, name='parser'),
     url(r'^man/$', views.man, name='man'),
+    url(r'^download/$', views.download, name='download'),
 ]